[pipeline] Refactor background tasks to dedicated module
Extract `BackgroundTask`, `BackgroundTaskFactory`,
`get_default_background_tasks`, and
`set_default_background_tasks` from
`_build.py` into a new `_bg_task.py` module.

This separates the background task abstractions
from the pipeline build logic,
making them independently importable.
